Moved index.js and index.test.js to "src" as cli-max.js and cli-max.test.js
