+++ src/cli-max.test.js
@@ -2,7 +2,7 @@
 
 describe('Module cli-max:', () => {
     test('should export an object with a "configure" method', () => {
-        const cli = require('./index');
+        const cli = require('./cli-max');
 
         expect(cli).toHaveProperty('configure');
         expect(cli.configure).toBeInstanceOf(Function);
